Revert "Made entry_ptr smart. (#16)"
This reverts commit 050ae8139e544e8a4ae6fbcce3a634114a4d7792.

Idea of this commit was to introduce smart pointers (make entry_ptr automatically clean it's
memory), but it added implementation overhead and removed some optimizations (like plain memcpy
or memmove of all node entry_ptrs).

We will use revert instead of reset, in order to have it in history if need it some day.